Fix 133090: MigrateActionTest#testMigrateUTCMany fails
In the file full of random UTC timestamps, there were duplicates - and these files were processed in file-on-disk order, which meant that even though they are sorted by name, you could end up accessing the duplicates in different orders (and while there were duplicates in the UTC, the file number is also in the label, which differs, causing unit test diffs and therefore a failure).

I processed the golden files a bit so there should be no duplicates now and the test should be stable.